services/memory/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1
sentence-transformers==2.7.0
numpy==1.24.3
python-multipart==0.0.6
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
PyJWT==2.8.0
structlog==24.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
